requirements.txt
defusedxml==0.4.1
Django==1.9.1
django-activity-stream==0.6.3
django-contrib-comments==1.7.3
django-filter==0.15.3
django-jsonfield==1.0.1
django-mobile==0.7.0
django-pagedown==0.1.1
django-taggit==0.21.3
django-tracking2==0.3.3
djangorestframework==3.5.3
markdown2==2.3.2
mysqlclient==1.3.7
oauthlib==2.0.1
psutil==5.0.0
PyJWT==1.4.2
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.10
requests==2.12.3
requests-oauthlib==0.7.0
shortuuid==0.4.3
six==1.10.0
uWSGI==2.0.13.1
